iio: adc: ad5592r_s : Fixed write, read functions and renamed them
-Renamed write function to spi_write and fixed the (8,0) mask to a (10,0)
one, now it works properly.
-Added a spi_nop function to use in the spi_read function(it writes a 0 tx,
and reads a value trough rx), it is used if we want to make a multi-channel
read where we read values while on the nop part is being transmitted, due
to the ADC being full-duplex.
-Renamed read function to spi_read and fixed it by using a single xfer,
first we transmit the readback value at the readback register, then
use the nop function to transmit 0 and also receive something if we
use the multi-channel functionality otherwise we can ignore the
received value, and finally we are receving the value through val,
and use the get_unaligned_be16 function.